+++ data.js
@@ -3,7 +3,8 @@ import path from 'path';
 export const data = {
   basedir: path.join('.', '.zap'),
   branch: path.join('.', '.zap', 'branch'),
-  defaultBranch: 'main'
+  defaultBranch: 'main',
+  version: '2.0.0',
 };
 
 export default data;